Update backup-azure-vms.md
Edit complete.

In the first figure, it would be better to spell out "VMs" as "virtual machines" if possible. Also, "Back up" should be all lowercase.

In other figures, please confirm that the names of virtual machines don't reveal any potentially sensitive information. 

The term "VM agent" was capitalized inconsistently. I went with "VM Agent" throughout, though I can't find an official standard on it. If you think "VM agent" makes more sense, please make the change throughout the article. (I left "agent" lowercase when it appears by itself.)

Procedures sometimes begin with an introductory sentence and sometimes begin with a procedure heading. It would be more consistent to standardize on a convention (whichever one you prefer).

On lines 110 and 116, please elaborate on "Once completed." It would be clearer to say "After <task> is completed" instead.

On line 138, please confirm that "unprotect" is a legitimate technical term. If not, it should be rephrased as something like "remove protection from."

On line 139, please check "VM config." It looks like this should be either "VM Agent configuration" or "virtual machine configuration" (for clarity).

On line 149, if "privileges" means "rights," "permissions," or "credentials," it would clearer to use one of those terms. 

The bulleted list that begins on line 160 should be a two-step procedure instead. Step 1 should be "Log on..." Step 2 should be "Right-click..." The other sentences shouldn't be formatted as separate steps because they're not really steps.

The term "IPs" appears a few times, beginning on line 174. Please make sure that this is clear enough and shouldn't be "IP ranges" or "IP addresses" instead.

On line 176, "whitelisted" is against Microsoft style. You might consider using something like "added to the list of safe IPs" unless "whitelisted" is more appropriate for the audience.

On line 177, I changed "PowerShell" to "Azure PowerShell." Please confirm that this change is accurate.

This article uses "boots up" a few times. Microsoft style generally prefers "starts" unless you're going for a casual tone.

Line 194 includes a four-step numbered list in a table. This content isn't really a procedure, so it looks like it should be converted to plain paragraph text.

Also on line 194, I rearranged and edited the text to avoid redundancy. Please make sure that the result is technically accurate and that the paragraph breaks are in the right place.